+++ bdc_oauth/blueprint.py
@@ -1,7 +1,7 @@
 from flask import Blueprint
 from flask_restplus import Api
 
-from api.status.controller import api as status_ns
+from bdc_oauth.status.controller import api as status_ns
 
 blueprint = Blueprint('oauth', __name__)
